resolved the is related issue

+++ src/cart/model.js
@@ -10,6 +10,7 @@
         userid) VALUES ($1,$2,$3,$4,$5,$6)
         ON CONFLICT (cartid) 
         DO UPDATE SET quantity = cart.quantity + EXCLUDED.quantity;`,
+        
         [cartId,cartName,price,carturl,quantity,userid])
          if(response.rowCount>0){
              return {error: false, data: response.rows , message: "cart created successfully"};
@@ -46,43 +47,56 @@
     }
  }
 
- const getCartData=async(req,client)=>{
-    try{
-        const response = await client.query(`SELECT * FROM "cart"`)
-        if(response){
-            return {error: false, data: response.rows , message: "cart read successfully"};
+const getCartData = async (req, client) => {
+    try {
+        const response = await client.query(`SELECT * FROM cart WHERE "userid" = $1`, [req.params.id])
+        if (response) {
+            return { error: false, data: response.rows, message: "cart read successfully" };
         } else {
-           return {error: true, message: "cart not read successfully"};
+            return { error: true, message: "cart not read successfully" };
         }
-    }  catch(err){
-        return {error: true, message: err.toString()};
+    } catch (err) {
+        return { error: true, message: err.toString() };
     }
- }
+}
 
- const deleteCartData=async(req,client)=>{
-        const response = await client.query(`DELETE FROM cart WHERE "cartid" = $1`,[req.params.id])
-        try{
-            if(response.rowCount>0){
-                return {error: false, message: "cart data deleted successfully"};
-            } else {
-               return {error: true, message: "cart data not deleted successfully"};
-            }
-            } catch(err){
-                return {error: true, message: err.toString()};
-            }
- }
- 
-const deleteCartAllData=async(req,client)=>{
-    const response = await client.query(`DELETE FROM cart`)
-    try{
-        if(response.rowCount>0){
-            return {error: false, message: "cart data deleted successfully"};
+const deleteCartData = async (req, client) => {
+    const response = await client.query(`DELETE FROM cart WHERE "cartid" = $1`, [req.params.id])
+    try {
+        if (response.rowCount > 0) {
+            return { error: false, message: "cart data deleted successfully" };
         } else {
-           return {error: true, message: "cart data not deleted successfully"};
+            return { error: true, message: "cart data not deleted successfully" };
         }
-        } catch(err){
-            return {error: true, message: err.toString()};
+    } catch (err) {
+        return { error: true, message: err.toString() };
+    }
+}
+
+const deleteCartAllData = async (req, client) => {
+    const response = await client.query(`DELETE FROM cart WHERE "userid" = $1`, [req.params.id])
+    try {
+        if (response.rowCount > 0) {
+            return { error: false, message: "cart data deleted successfully" };
+        } else {
+            return { error: true, message: "cart data not deleted successfully" };
         }
+    } catch (err) {
+        return { error: true, message: err.toString() };
+    }
+}
+
+const getCartDetail = async (req, client) => {
+    try {
+        const response = await client.query(`SELECT * FROM cart WHERE "cartid" = $1`, [req.params.id])
+        if (response) {
+            return { error: false, data: response.rows, message: "cart read successfully" };
+        } else {
+            return { error: true, message: "cart not read successfully" };
+        }
+    } catch (err) {
+        return { error: true, message: err.toString() };
+    }
 }
 
 export {
@@ -90,5 +104,6 @@ export {
     getCartData,
     deleteCartData,
     decCart,
-    deleteCartAllData
- }
+    deleteCartAllData,
+    getCartDetail
+}
